@@ -32,12 +24,19 @@
         </developer>
     </developers>
 
+    <scm>
+        <connection>scm:git:https://github.com/unbearables/graylog-plugin-jira.git</connection>
+        <developerConnection>scm:git:https://github.com/unbearables</developerConnection>
+        <url>https://github.com/unbearables/graylog-plugin-jira</url>
+        <tag>v2.0.0</tag>
+    </scm>
+
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.release>17</maven.compiler.release>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <graylog.version>${project.parent.version}</graylog.version> <!-- The dependent graylog2-server lib version -->
-        <graylog.version.required>3.1.3</graylog.version.required> <!-- The minimum graylog version required to run the plugin -->
 
         <!-- Plugins will not be deployed by default - set to `false` if you actually want to deploy it -->
         <maven.source.skip>true</maven.source.skip>
@@ -86,13 +177,51 @@
 
     <build>
         <resources>
-            <resource><directory>${web.build-dir}</directory></resource>
+            <resource><directory>build</directory></resource>
             <resource>
-              <directory>src/main/resources</directory>
-              <filtering>true</filtering>
+                <directory>src/main/resources</directory>
+                <filtering>true</filtering>
             </resource>
         </resources>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <release>17</release>
+                    <showWarnings>true</showWarnings>
+                    <showDeprecation>true</showDeprecation>
+                    <forceLegacyJavacApi>true</forceLegacyJavacApi>
+                    <compilerArgs>
+                        <arg>-XDcompilePolicy=simple</arg>
+                        <arg>--should-stop=ifError=FLOW</arg>
+                        <arg>-Xplugin:ErrorProne</arg>
+                    </compilerArgs>
+                    <annotationProcessorPaths>
+                        <path>
+                            <groupId>com.google.errorprone</groupId>
+                            <artifactId>error_prone_core</artifactId>
+                            <version>${error-prone.version}</version>
+                        </path>
+                        <path>
+                            <groupId>com.google.auto.service</groupId>
+                            <artifactId>auto-service</artifactId>
+                            <version>${auto-service.version}</version>
+                        </path>
+                        <path>
+                            <groupId>com.google.auto.value</groupId>
+                            <artifactId>auto-value</artifactId>
+                            <version>${auto-value.version}</version>
+                        </path>
+                        <path>
+                            <groupId>org.graylog.autovalue</groupId>
+                            <artifactId>auto-value-javabean</artifactId>
+                            <version>${auto-value-javabean.version}</version>
+                        </path>
+                    </annotationProcessorPaths>
+                </configuration>
+            </plugin>
+
             <plugin>
                 <artifactId>maven-assembly-plugin</artifactId>
                 <configuration>
